refactor: migrate nested model arguments to TypeScript
